+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#24](https://github.com/LaxarJS/laxar-angular-adapter/issues/24): adjusted to adapter API changes (laxar#358)
 - [#23](https://github.com/LaxarJS/laxar-angular-adapter/issues/23): fixed testability and multi-instance operation
 - [#22](https://github.com/LaxarJS/laxar-angular-adapter/issues/22): removed AngularJS specific `axI18n` injection in favor of widget-services version
 - [#21](https://github.com/LaxarJS/laxar-angular-adapter/issues/21): fixed bower.json `main` entry